MapDto's besser umbenannt, GeoCoding Klasse verbessert, DataIntializerBeispiele eingefügt, Fehler in Artikelerstellung.html gefunden und detailansicht.html fertiggestellt (Lisch195)

+++ Elite-Verleih/src/main/java/de/hhu/abschlussprojektmeangirlzhackstreetboys/eliteverleih/modell/Artikel.java
@@ -28,8 +28,8 @@ public class Artikel {
 
     private String artikelOrt;
 
-    private int artikelOrtX;
-    private int artikelOrtY;
+    private double artikelOrtX;
+    private double artikelOrtY;
 
     @OneToMany(cascade = CascadeType.ALL)
     private List<Ausleihe> ausgeliehen;
